refactor: move tests into named folders
