+++ test/mutiple-client/non-optimistic.test.js
@@ -47,9 +47,9 @@ afterAll(() => {
 
 test('Multiple clients upload js feeds to build server', async () => {
     expect.assertions(1);
-    const js1 = resolve('../assets/a.js');
-    const js2 = resolve('../assets/b.js');
-    const js3 = resolve('../assets/c.js');
+    const js1 = resolve('../../assets/a.js');
+    const js2 = resolve('../../assets/b.js');
+    const js3 = resolve('../../assets/c.js');
     const result = await Promise.all([
         client.uploadFeed([js1]),
         client.uploadFeed([js2]),
@@ -60,9 +60,9 @@ test('Multiple clients upload js feeds to build server', async () => {
 
 test('Multiple clients upload css feeds to build server', async () => {
     expect.assertions(1);
-    const css1 = resolve('../assets/a.css');
-    const css2 = resolve('../assets/b.css');
-    const css3 = resolve('../assets/c.css');
+    const css1 = resolve('../../assets/a.css');
+    const css2 = resolve('../../assets/b.css');
+    const css3 = resolve('../../assets/c.css');
     const result = await Promise.all([
         client.uploadFeed([css1]),
         client.uploadFeed([css2]),
@@ -73,9 +73,9 @@ test('Multiple clients upload css feeds to build server', async () => {
 
 test('Multiple clients get uploaded js feeds from build server', async () => {
     expect.assertions(1);
-    const js1 = resolve('../assets/a.js');
-    const js2 = resolve('../assets/b.js');
-    const js3 = resolve('../assets/c.js');
+    const js1 = resolve('../../assets/a.js');
+    const js2 = resolve('../../assets/b.js');
+    const js3 = resolve('../../assets/c.js');
     const [{ file: f1 }, { file: f2 }, { file: f3 }] = await Promise.all([
         client.uploadFeed([js1]),
         client.uploadFeed([js2]),
@@ -91,9 +91,9 @@ test('Multiple clients get uploaded js feeds from build server', async () => {
 
 test('Multiple clients get uploaded css feeds from build server', async () => {
     expect.assertions(1);
-    const css1 = resolve('../assets/a.css');
-    const css2 = resolve('../assets/b.css');
-    const css3 = resolve('../assets/c.css');
+    const css1 = resolve('../../assets/a.css');
+    const css2 = resolve('../../assets/b.css');
+    const css3 = resolve('../../assets/c.css');
     const [{ file: f1 }, { file: f2 }, { file: f3 }] = await Promise.all([
         client.uploadFeed([css1]),
         client.uploadFeed([css2]),
@@ -109,9 +109,9 @@ test('Multiple clients get uploaded css feeds from build server', async () => {
 
 test('Client requests bundling of a js feed on build server', async () => {
     expect.assertions(1);
-    const js1 = resolve('../assets/a.js');
-    const js2 = resolve('../assets/b.js');
-    const js3 = resolve('../assets/c.js');
+    const js1 = resolve('../../assets/a.js');
+    const js2 = resolve('../../assets/b.js');
+    const js3 = resolve('../../assets/c.js');
     const [{ file: f1 }, { file: f2 }, { file: f3 }] = await Promise.all([
         client.uploadFeed([js1]),
         client.uploadFeed([js2]),
@@ -123,9 +123,9 @@ test('Client requests bundling of a js feed on build server', async () => {
 
 test('Client requests bundling of a css feed on build server', async () => {
     expect.assertions(1);
-    const css1 = resolve('../assets/a.css');
-    const css2 = resolve('../assets/b.css');
-    const css3 = resolve('../assets/c.css');
+    const css1 = resolve('../../assets/a.css');
+    const css2 = resolve('../../assets/b.css');
+    const css3 = resolve('../../assets/c.css');
     const [{ file: f1 }, { file: f2 }, { file: f3 }] = await Promise.all([
         client.uploadFeed([css1]),
         client.uploadFeed([css2]),
@@ -137,9 +137,9 @@ test('Client requests bundling of a css feed on build server', async () => {
 
 test('Client gets bundled js file from build server', async () => {
     expect.assertions(1);
-    const js1 = resolve('../assets/a.js');
-    const js2 = resolve('../assets/b.js');
-    const js3 = resolve('../assets/c.js');
+    const js1 = resolve('../../assets/a.js');
+    const js2 = resolve('../../assets/b.js');
+    const js3 = resolve('../../assets/c.js');
     const [{ file: f1 }, { file: f2 }, { file: f3 }] = await Promise.all([
         client.uploadFeed([js1]),
         client.uploadFeed([js2]),
@@ -154,9 +154,9 @@ test('Client gets bundled js file from build server', async () => {
 
 test('Client gets bundled css file from build server', async () => {
     expect.assertions(1);
-    const css1 = resolve('../assets/a.css');
-    const css2 = resolve('../assets/b.css');
-    const css3 = resolve('../assets/c.css');
+    const css1 = resolve('../../assets/a.css');
+    const css2 = resolve('../../assets/b.css');
+    const css3 = resolve('../../assets/c.css');
     const [{ file: f1 }, { file: f2 }, { file: f3 }] = await Promise.all([
         client.uploadFeed([css1]),
         client.uploadFeed([css2]),
@@ -171,9 +171,9 @@ test('Client gets bundled css file from build server', async () => {
 
 test('Bundled js feeds from build server are deduped', async () => {
     expect.assertions(1);
-    const js1 = resolve('../assets/duped-before-uploading.js');
-    const js2 = resolve('../assets/duped-before-uploading-2.js');
-    const js3 = resolve('../assets/duped-before-uploading-3.js');
+    const js1 = resolve('../../assets/duped-before-uploading.js');
+    const js2 = resolve('../../assets/duped-before-uploading-2.js');
+    const js3 = resolve('../../assets/duped-before-uploading-3.js');
     const uploadResponse1 = await client.uploadFeed([js1]);
     const uploadResponse2 = await client.uploadFeed([js2]);
     const uploadResponse3 = await client.uploadFeed([js3]);
