+++ test/single-client/non-optimistic.test.js
@@ -48,13 +48,13 @@ afterAll(() => stopServer());
 
 test('Client uploads a js feed to build server', async () => {
     expect.assertions(1);
-    const result = await client.uploadFeed([resolve('../assets/index.js')]);
+    const result = await client.uploadFeed([resolve('../../assets/index.js')]);
     expect(result).toMatchSnapshot();
 });
 
 test('Client gets uploaded js feed from build server', async () => {
     expect.assertions(1);
-    const result = await client.uploadFeed([resolve('../assets/index.js')]);
+    const result = await client.uploadFeed([resolve('../../assets/index.js')]);
     const { body } = await request
         .get(`/feed/${result.file}`)
         .set('Accept', 'application/json')
@@ -65,13 +65,13 @@ test('Client gets uploaded js feed from build server', async () => {
 
 test('Client uploads a css feed to build server', async () => {
     expect.assertions(1);
-    const result = await client.uploadFeed([resolve('../assets/style.css')]);
+    const result = await client.uploadFeed([resolve('../../assets/style.css')]);
     expect(result).toMatchSnapshot();
 });
 
 test('Client gets uploaded css feed from build server', async () => {
     expect.assertions(1);
-    const result = await client.uploadFeed([resolve('../assets/style.css')]);
+    const result = await client.uploadFeed([resolve('../../assets/style.css')]);
     const { body } = await request
         .get(`/feed/${result.file}`)
         .set('Accept', 'application/json')
@@ -83,7 +83,7 @@ test('Client gets uploaded css feed from build server', async () => {
 test('Client requests bundling of a js feed on build server', async () => {
     expect.assertions(1);
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/index.js'),
+        resolve('../../assets/index.js'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
@@ -95,7 +95,7 @@ test('Client requests bundling of a js feed on build server', async () => {
 test('Client gets bundled js file from build server', async () => {
     expect.assertions(1);
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/index.js'),
+        resolve('../../assets/index.js'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
@@ -110,7 +110,7 @@ test('Client gets bundled js file from build server', async () => {
 test('Client requests bundling of a css feed on build server', async () => {
     expect.assertions(1);
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/style.css'),
+        resolve('../../assets/style.css'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
@@ -122,7 +122,7 @@ test('Client requests bundling of a css feed on build server', async () => {
 test('Client gets bundled css file from build server', async () => {
     expect.assertions(1);
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/style.css'),
+        resolve('../../assets/style.css'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
@@ -137,7 +137,7 @@ test('Client gets bundled css file from build server', async () => {
 test('Uploaded js feed from build server is not deduped', async () => {
     expect.assertions(1);
     const result = await client.uploadFeed([
-        resolve('../assets/duped-before-uploading.js'),
+        resolve('../../assets/duped-before-uploading.js'),
     ]);
     const { body } = await request
         .get(`/feed/${result.file}`)
@@ -150,7 +150,7 @@ test('Uploaded js feed from build server is not deduped', async () => {
 test('Bundled js feed from build server is deduped', async () => {
     expect.assertions(1);
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/duped-before-uploading.js'),
+        resolve('../../assets/duped-before-uploading.js'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
@@ -165,7 +165,7 @@ test('Bundled js feed from build server is deduped', async () => {
 test('NODE_ENV variables replaced with hard coded values', async () => {
     expect.assertions(5);
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/node_envs.js'),
+        resolve('../../assets/node_envs.js'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
@@ -187,7 +187,7 @@ test('NODE_ENV variables trigger dead code elimination (DCE)', async () => {
     await startServer('production');
 
     const uploadResponse = await client.uploadFeed([
-        resolve('../assets/node_envs.js'),
+        resolve('../../assets/node_envs.js'),
     ]);
     const bundleResponse = await client.createRemoteBundle(
         [uploadResponse.file],
